done lastMessage issue

+++ src/api/graphql/resolvers/mutations/DeleteMessage.js
@@ -1,22 +1,27 @@
-const DeleteMessage = async ({ args, pubsub, Message }) => {
+const DeleteMessage = async ({ args, pubsub, Message, Conversation }) => {
     try {
         const { id, senderId, conversationId } = args
+
+        const isConversation = await Conversation.findById(conversationId)
+        if (!isConversation) return { success: false, message: "Conversation doesn't exist!" }
 
         const isMessage = await Message.deleteOne({ _id: id, senderId, conversationId })
         if (!isMessage) return { success: false, message: "Message does not exist." }
 
         const messages = await Message.find({ conversationId }).sort({ createdAt: -1 }).limit(50)
+
+        isConversation.lastMessage = messages[0]?.content
+        isConversation.lastMessageType = messages[0]?.type
+        await isConversation.save()
+
         pubsub.publish(
             conversationId,
             { messageAdded: { success: true, message: "", messages: messages.reverse() } }
         );
-        pubsub.publish(
-            `${conversationId}LM`,
-            {
-                lastMessageAdded:
-                    { success: true, message: "", messages: messages.length ? messages[messages.length - 1] : null }
-            }
-        );
+        isConversation.members.forEach(async (id) => {
+            const conversations = await Conversation.find({ members: { $in: [id] } }).sort({ updatedAt: -1 })
+            pubsub.publish(id, { conversationAdded: { success: true, message: "", conversations } });
+        })
 
         return { success: true, message: "Message has been deleted." }
     } catch (err) {
